+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#2259](https://github.com/oshi/oshi/pull/2259): Cache AIX partition list to reduce disk reads from lspv - [@dbwiddis](https://github.com/dbwiddis).
 * [#2260](https://github.com/oshi/oshi/pull/2260): Use regex to pre-filter to parseable CPU numbers for ARM Macs - [@dbwiddis](https://github.com/dbwiddis).
 * [#2262](https://github.com/oshi/oshi/pull/2262): Consistent treatment of AIX tick lengths - [@dbwiddis](https://github.com/dbwiddis).
+* [#2263](https://github.com/oshi/oshi/pull/2263): Don't assume ticks match logical processor count - [@dbwiddis](https://github.com/dbwiddis).
 
 # 6.3.0 (2022-10-16), 6.3.1 (2022-10-30), 6.3.2 (2022-11-16)